AUTH SETUP

+++ src/tasks/tasks.service.ts
@@ -7,6 +7,7 @@ import { TaskRepository } from "./task.repository";
 import { InjectRepository } from "@nestjs/typeorm";
 import { Task } from "./task.entity";
 import { Repository } from "typeorm";
+import { User } from "src/auth/user.entity";
 
 @Injectable()
 export class TasksService {
@@ -15,31 +16,37 @@ export class TasksService {
 		private taskRepository: TaskRepository,
 	) {}
 
-	async getTasks(filterDto: GetTasksFilterDto): Promise<Task[]> {
-		return await this.taskRepository.getTasks(filterDto);
+	async getTasks(filterDto: GetTasksFilterDto, user: User): Promise<Task[]> {
+		return await this.taskRepository.getTasks(filterDto, user);
 	}
 
-	async getTaskById(id: number): Promise<Task> {
-		const task = await this.taskRepository.findOneBy({ id });
+	async getTaskById(id: number, user: User): Promise<Task> {
+		const task = await this.taskRepository.findOne({
+			where: { id, userId: user.id },
+		});
 		if (!task) {
 			throw new NotFoundException();
 		}
 		return task;
 	}
 
-	async deleteTaskById(id: number): Promise<void> {
-		const result = await this.taskRepository.delete(id);
+	async deleteTaskById(id: number, user: User): Promise<void> {
+		const result = await this.taskRepository.delete({ id, userId: user.id });
 		if (!result.affected) {
-			throw new NotFoundException();
+			throw new NotFoundException(`Task with ID '${id}' not found`);
 		}
 	}
 
-	async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-		return this.taskRepository.createTask(createTaskDto);
+	async createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
+		return this.taskRepository.createTask(createTaskDto, user);
 	}
 
-	async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-		const task = await this.getTaskById(id);
+	async updateTaskStatus(
+		id: number,
+		status: TaskStatus,
+		user: User,
+	): Promise<Task> {
+		const task = await this.getTaskById(id, user);
 		task.status = status;
 		await task.save();
 		return task;
